+++ src/RequestBuilder/RequestBuilderFactory.php
@@ -61,6 +61,11 @@ class RequestBuilderFactory
         return $this->createConfiguredBuilder(RecommendationRequestBuilder::class, $recommendation);
     }
 
+    public function forget(): ForgetRequestBuilder
+    {
+        return $this->createConfiguredBuilder(ForgetRequestBuilder::class);
+    }
+
     public function resetDatabase(): ResetDatabaseRequestBuilder
     {
         return $this->createConfiguredBuilder(ResetDatabaseRequestBuilder::class);