+++ tests/unit/RequestBuilder/RequestBuilderFactoryTest.php
@@ -6,6 +6,7 @@ use Lmc\Matej\Http\RequestManager;
 use Lmc\Matej\Model\Command\ItemProperty;
 use Lmc\Matej\Model\Command\ItemPropertySetup;
 use Lmc\Matej\Model\Command\Sorting;
+use Lmc\Matej\Model\Command\UserForget;
 use Lmc\Matej\Model\Command\UserRecommendation;
 use Lmc\Matej\Model\Request;
 use Lmc\Matej\Model\Response;
@@ -64,6 +65,10 @@ class RequestBuilderFactoryTest extends TestCase
             $builder->addSorting(Sorting::create('item-id', ['item1', 'item2']));
         };
 
+        $forgetInit = function (ForgetRequestBuilder $builder): void {
+            $builder->addUser(UserForget::anonymize('test-user-for-anonymization'));
+        };
+
         $voidInit = function ($builder): void {};
 
         $userRecommendation = UserRecommendation::create('user-id', 1, 'test-scenario', 0.5, 3600);
@@ -76,6 +81,7 @@ class RequestBuilderFactoryTest extends TestCase
             ['campaign', CampaignRequestBuilder::class, $campaignInit],
             ['sorting', SortingRequestBuilder::class, $voidInit, Sorting::create('user-a', ['item-a', 'item-b', 'item-c'])],
             ['recommendation', RecommendationRequestBuilder::class, $voidInit, $userRecommendation],
+            ['forget', ForgetRequestBuilder::class, $forgetInit],
             ['resetDatabase', ResetDatabaseRequestBuilder::class, $voidInit],
         ];
     }